Adds relative root dir helper, reorganizes folders
