+++ package.json
@@ -7,6 +7,7 @@
     "gatsby-image": "^2.0.13",
     "gatsby-plugin-layout": "^1.0.12",
     "gatsby-plugin-react-helmet": "^3.0.0",
+    "gatsby-plugin-root-import": "^2.0.5",
     "gatsby-plugin-sass": "^2.0.10",
     "gatsby-plugin-sharp": "^2.0.6",
     "gatsby-plugin-stripe": "^1.2.0",
